Prepared for endian conversion
git-svn-id: https://svn.nfit.au.dk/svn/beta/trunk@11051 678b6752-a8a0-49ff-a511-7c2a3241fcd4

+++ betarun/P/des.h
@@ -11,7 +11,7 @@
  */
 
 typedef struct areaTableEntry {
-  short allocated;
+  u_long allocated;
   u_long storageOffset;
   u_long areaTop;
   u_long areaSize;
